reorder workflow utils, fix tests/imports
