+++ src/dolphin/workflows/__init__.py
@@ -1,4 +1,4 @@
 """Tools for creating workflows and running displacement workflows."""
 from ._enums import *  # noqa
-from ._grouping import *  # noqa
+from ._utils import *  # noqa
 from .config import *  # noqa
