--- src/dolphin/workflows/_grouping.py
@@ -1,82 +1,0 @@
-import itertools
-import re
-from pathlib import Path
-from typing import Dict, List, Pattern, Sequence, Union
-
-from dolphin._log import get_log
-from dolphin._types import Filename
-
-from .config import OPERA_BURST_RE
-
-logger = get_log(__name__)
-
-__all__ = ["group_by_burst"]
-
-
-def group_by_burst(
-    file_list: Sequence[Filename],
-    burst_id_fmt: Union[str, Pattern[str]] = OPERA_BURST_RE,
-    minimum_slcs: int = 2,
-) -> Dict[str, List[Path]]:
-    """Group Sentinel CSLC files by burst.
-
-    Parameters
-    ----------
-    file_list: List[Filename]
-        path to folder containing CSLC files
-    burst_id_fmt: str
-        format of the burst id in the filename.
-        Default is [OPERA_BURST_RE][]
-    minimum_slcs: int
-        Minimum number of SLCs needed to run the workflow for each burst.
-        If there are fewer SLCs in a burst, it will be skipped and
-        a warning will be logged.
-
-    Returns
-    -------
-    dict
-        key is the burst id of the SLC acquisition
-        Value is a list of Paths on that burst:
-        {
-            't087_185678_iw2': [Path(...), Path(...),],
-            't087_185678_iw3': [Path(...),... ],
-        }
-    """
-
-    def get_burst_id(filename):
-        m = re.search(burst_id_fmt, str(filename))
-        if not m:
-            raise ValueError(f"Could not parse burst id from {filename}")
-        return m.group()
-
-    def sort_by_burst_id(file_list):
-        """Sort files by burst id."""
-        burst_ids = [get_burst_id(f) for f in file_list]
-        file_burst_tups = sorted(
-            [(Path(f), d) for f, d in zip(file_list, burst_ids)],
-            # use the date or dates as the key
-            key=lambda f_d_tuple: f_d_tuple[1],  # type: ignore
-        )
-        # Unpack the sorted pairs with new sorted values
-        file_list, burst_ids = zip(*file_burst_tups)  # type: ignore
-        return file_list
-
-    sorted_file_list = sort_by_burst_id(file_list)
-    # Now collapse into groups, sorted by the burst_id
-    grouped_images = {
-        burst_id: list(g)
-        for burst_id, g in itertools.groupby(
-            sorted_file_list, key=lambda x: get_burst_id(x)
-        )
-    }
-    # Make sure that each burst has at least the minimum number of SLCs
-    out = {}
-    for burst_id, slc_list in grouped_images.items():
-        if len(slc_list) < minimum_slcs:
-            logger.warning(
-                f"Skipping burst {burst_id} because it has only {len(slc_list)} SLCs."
-                f"Minimum number of SLCs is {minimum_slcs}"
-            )
-        else:
-            out[burst_id] = slc_list
-    return out
